Use load-grunt-config to make Grunt more maintainable
