+++ templates/common/root/grunt/_cdnify.js
@@ -1,0 +1,6 @@
+// Replace Google CDN references
+module.exports = {
+  dist: {
+    html: ['<%%= yeoman.dist %>/*.html']
+  }
+};
