+++ templates/common/root/grunt/_clean.js
@@ -1,0 +1,14 @@
+// Empties folders to start fresh
+module.exports = {
+  dist: {
+    files: [{
+      dot: true,
+      src: [
+        '.tmp',
+        '<%%= yeoman.dist %>/{,*/}*',
+        '!<%%= yeoman.dist %>/.git*'
+      ]
+    }]
+  },
+  server: '.tmp'
+};
