+++ templates/common/root/grunt/_cssmin.js
@@ -1,0 +1,14 @@
+// The following *-min task will produce minified files in the dist folder
+// By default, your `index.html`'s <!-- Usemin block --> will take care of
+// minification. These next options are pre-configured if you do not wish
+// to use the Usemin blocks.
+// Also make sure do uncomment the configuration in uglify.js and concat.js
+module.exports = {
+//   dist: {
+//     files: {
+//       '<%%= yeoman.dist %>/styles/main.css': [
+//         '.tmp/styles/{,*/}*.css'
+//       ]
+//     }
+//   }
+};
