+++ templates/common/root/grunt/_filerev.js
@@ -1,0 +1,11 @@
+// Renames files for browser caching purposes
+module.exports = {
+  dist: {
+    src: [
+      '<%%= yeoman.dist %>/scripts/{,*/}*.js',
+      '<%%= yeoman.dist %>/styles/{,*/}*.css',
+      '<%%= yeoman.dist %>/images/{,*/}*.{png,jpg,jpeg,gif,webp,svg}',
+      '<%%= yeoman.dist %>/styles/fonts/*'
+    ]
+  }
+};
