+++ templates/common/root/grunt/_htmlmin.js
@@ -1,0 +1,17 @@
+module.exports = {
+  dist: {
+    options: {
+      collapseWhitespace: true,
+      conservativeCollapse: true,
+      collapseBooleanAttributes: true,
+      removeCommentsFromCDATA: true,
+      removeOptionalTags: true
+    },
+    files: [{
+      expand: true,
+      cwd: '<%%= yeoman.dist %>',
+      src: ['*.html', 'views/{,*/}*.html'],
+      dest: '<%%= yeoman.dist %>'
+    }]
+  }
+};
