+++ templates/common/root/grunt/_imagemin.js
@@ -1,0 +1,10 @@
+module.exports = {
+  dist: {
+    files: [{
+      expand: true,
+      cwd: '<%%= yeoman.app %>/images',
+      src: '{,*/}*.{png,jpg,jpeg,gif}',
+      dest: '<%%= yeoman.dist %>/images'
+    }]
+  }
+};
