+++ templates/common/root/grunt/_karma.js
@@ -1,0 +1,10 @@
+// Test settings
+module.exports = {
+  unit: {
+    configFile: 'test/karma.conf.<% if (coffee) {
+      %>coffee<% } else {
+      %>js<% }
+      %>',
+    singleRun: true
+  }
+};
