+++ templates/common/root/grunt/_svgmin.js
@@ -1,0 +1,10 @@
+module.exports = {
+  dist: {
+    files: [{
+      expand: true,
+      cwd: '<%%= yeoman.app %>/images',
+      src: '{,*/}*.svg',
+      dest: '<%%= yeoman.dist %>/images'
+    }]
+  }
+};
