+++ templates/common/root/grunt/tasks/_default.js
@@ -1,0 +1,7 @@
+module.exports = function (grunt) {
+  grunt.registerTask('default', [
+    'newer:jshint',
+    'test',
+    'build'
+  ]);
+};
