+++ templates/common/root/grunt/tasks/_server.js
@@ -1,0 +1,6 @@
+module.exports = function (grunt) {
+  grunt.registerTask('server', 'DEPRECATED TASK. Use the "serve" task instead', function (target) {
+    grunt.log.warn('The `server` task has been deprecated. Use `grunt serve` to start a server.');
+    grunt.task.run(['serve:' + target]);
+  });
+};
