+++ templates/common/root/grunt/tasks/_test.js
@@ -1,0 +1,9 @@
+module.exports = function (grunt) {
+  grunt.registerTask('test', [
+    'clean:server',
+    'concurrent:test',
+    'autoprefixer',
+    'connect:test',
+    'karma'
+  ]);
+};
